Combine event and event register
